add entities and change mongo entity location
